requirements-windows.txt
requests>=2.31,<3
pyinstaller>=6,<7
winrt-runtime>=3.2.1,<4
winrt-Windows.Foundation>=3.2.1,<4
winrt-Windows.Foundation.Collections>=3.2.1,<4
winrt-Windows.Media.Core>=3.2.1,<4
winrt-Windows.Media.Playback>=3.2.1,<4
winrt-Windows.Media.SpeechSynthesis>=3.2.1,<4
winrt-Windows.Storage.Streams>=3.2.1,<4
